use enum instead of strings for authn scheme names
