+++ nexus/auth/src/authn/external/session_cookie.rs
@@ -58,7 +58,7 @@ pub trait SessionStore {
 // generic cookie name is recommended by OWASP
 pub const SESSION_COOKIE_COOKIE_NAME: &str = "session";
 pub const SESSION_COOKIE_SCHEME_NAME: authn::SchemeName =
-    authn::SchemeName("session_cookie");
+    authn::SchemeName::SessionCookie;
 
 /// Generate session cookie header
 pub fn session_cookie_header_value(
