+++ nexus/auth/src/lib.rs
@@ -3,9 +3,6 @@ pub mod authz;
 pub mod context;
 pub mod storage;
 
-#[macro_use]
-extern crate newtype_derive;
-
 #[allow(unused_imports)]
 #[macro_use]
 extern crate slog;
